tests/wpt/web-platform-tests/css/css-grid/layout-algorithm/grid-content-distribution-must-account-for-track-sizing-004.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Grid Layout Test: Content Distribution and the track sizing algorithm</title>
<link rel="author" title="Javier Fernandez Garcia-Boente" href="mailto:jfernandez@igalia.com">
<link rel="help" href="https://drafts.csswg.org/css-grid/#grid-align">
<link rel="help" href="https://drafts.csswg.org/css-align/#content-distribution">
<link rel="help" href="https://drafts.csswg.org/css-align/#valdef-align-content-space-between">
<link rel="stylesheet" href="../../support/alignment.css">
<meta name="flags" content="ahem">
<meta name="assert" content="Content Distribution offset doesn't account for relative sized tracks with non-spanning items.">
<link rel="stylesheet" type="text/css" href="/fonts/ahem.css" />
<style>
.grid {
  display: inline-grid;
  background: grey;
  grid-template-columns: 50%;
  font: 20px/1 Ahem;
}

.item1 {
  background: green;
  grid-column: 1;
  grid-row: 1;
}
.item2 {
  background: blue;
  grid-column: 2;
  grid-row: 1;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>

<body onload="checkLayout('.grid')">
    <div class="grid justifyContentSpaceBetween" data-expected-width="220" data-expected-height="40">
        <div class="item1" data-expected-width="110" data-expected-height="40">XXXX XXX</div>
        <div class="item2" data-expected-width="60" data-expected-height="40">XXX</div>
    </div>
</body>
